working version with diff + fill logic

+++ cmd/cli/main.go
@@ -19,8 +19,8 @@ func parseFlags() ([]string, bool) {
 	return repos, *force
 }
 
-func commitCompletedCallback(curr, total int) {
-	log.Info().Int("current", curr).Int("total", total).Msg("Processed commits")
+func commitCompletedCallback(curr, total uint64) {
+	log.Info().Uint64("current", curr).Uint64("total", total).Msg("Processed commits")
 }
 
 func main() {
